IMAP Improvements (#558)
* Fixing an issue where scrollviewer overrides a part of template in mail list. Adjusted zoomed out header grid's corner radius.

* IDLE implementation, imap synchronization strategies basics and condstore synchronization.

* Adding iCloud and Yahoo as special IMAP handling scenario.

* iCloud special imap handling.

* Support for killing synchronizers.

* Update privacy policy url.

* Batching condstore downloads into 50, using SORT extension for searches if supported.

* Bumping some nugets. More on the imap synchronizers.

* Delegating idle synchronizations to server to post-sync operations.

* Update mailkit to resolve qresync bug with iCloud.

* Fixing remote highest mode seq checks for qresync and condstore synchronizers.

* Yahoo custom settings.

* Bump google sdk package.

* Fixing the build issue....

* NRE on canceled token accounts during setup.

* Server crash handlers.

* Remove ARM32. Upgrade server to .NET 9.

* Fix icons for yahoo and apple.

* Fixed an issue where disabled folders causing an exception on forced sync.

* Remove smtp encoding constraint.

* Remove commented code.

* Fixing merge conflict

* Addressing double registrations for mailkit remote folder events in synchronizers.

* Making sure idle canceled result is not reported.

* Fixing custom imap server dialog opening.

* Fixing the issue with account creation making the previously selected account as selected as well.

* Fixing app close behavior and logging app close.

+++ Wino.Mail.ViewModels/AccountDetailsPageViewModel.cs
@@ -12,6 +12,7 @@ using Wino.Core.Domain.Interfaces;
 using Wino.Core.Domain.Models.Folders;
 using Wino.Core.Domain.Models.Navigation;
 using Wino.Messaging.Client.Navigation;
+using Wino.Messaging.Server;
 using Wino.Messaging.UI;
 
 namespace Wino.Mail.ViewModels
@@ -20,6 +21,7 @@ namespace Wino.Mail.ViewModels
     {
         private readonly IMailDialogService _dialogService;
         private readonly IAccountService _accountService;
+        private readonly IWinoServerConnectionManager _serverConnectionManager;
         private readonly IFolderService _folderService;
 
         public MailAccount Account { get; set; }
@@ -48,10 +50,12 @@ namespace Wino.Mail.ViewModels
 
         public AccountDetailsPageViewModel(IMailDialogService dialogService,
             IAccountService accountService,
+            IWinoServerConnectionManager serverConnectionManager,
             IFolderService folderService)
         {
             _dialogService = dialogService;
             _accountService = accountService;
+            _serverConnectionManager = serverConnectionManager;
             _folderService = folderService;
         }
 
@@ -102,13 +106,17 @@ namespace Wino.Mail.ViewModels
             if (!confirmation)
                 return;
 
-            await _accountService.DeleteAccountAsync(Account);
 
-            // TODO: Server: Cancel ongoing calls from server for this account.
+            var isSynchronizerKilledResponse = await _serverConnectionManager.GetResponseAsync<bool, KillAccountSynchronizerRequested>(new KillAccountSynchronizerRequested(Account.Id));
 
-            _dialogService.InfoBarMessage(Translator.Info_AccountDeletedTitle, string.Format(Translator.Info_AccountDeletedMessage, Account.Name), InfoBarMessageType.Success);
+            if (isSynchronizerKilledResponse.IsSuccess)
+            {
+                await _accountService.DeleteAccountAsync(Account);
 
-            Messenger.Send(new BackBreadcrumNavigationRequested());
+                _dialogService.InfoBarMessage(Translator.Info_AccountDeletedTitle, string.Format(Translator.Info_AccountDeletedMessage, Account.Name), InfoBarMessageType.Success);
+
+                Messenger.Send(new BackBreadcrumNavigationRequested());
+            }
         }
 
         public override async void OnNavigatedTo(NavigationMode mode, object parameters)
